Add func tests for retrying Cucumber tests using TestNG runner
